+++ inc/model/MediaOpaque.h
@@ -23,7 +23,7 @@ private:
     void *m_pXmscCodec = NULL;
 
 private:
-	int _openFile(const wstring& strFile, bool bXmsc);
+	long long _openFile(const wstring& strFile, bool bXmsc);
 
 	UINT _checkDuration()
 	{
@@ -32,15 +32,15 @@ private:
 		return uDuration;
 	}
 
-	static UINT _checkDuration(const wstring& strFile, bool bXmsc, int& nFileSize);
+	static UINT _checkDuration(const wstring& strFile, bool bXmsc, long long& nFileSize);
 
 public:
-    int openFile(const wstring& strFile, bool bXmsc)
+    long long openFile(const wstring& strFile, bool bXmsc)
 	{
 		return _openFile(strFile, bXmsc);
     }
 
-	int openFile(IMedia& media)
+	long long openFile(IMedia& media)
 	{
 		return _openFile(media.GetAbsPath(), media.isXmsc());
 	}
@@ -58,7 +58,7 @@ public:
     }
 
 #else
-	UINT checkFileDuration(const wstring& strFile, int& nFileSize)
+	UINT checkFileDuration(const wstring& strFile, long long& nFileSize)
 	{
 		nFileSize = _openFile(strFile, false);
 		if (nFileSize <= 0)
@@ -71,13 +71,13 @@ public:
 
 	UINT checkFileDuration(const wstring& strFile)
 	{
-		int nFileSize = 0;
+		long long nFileSize = 0;
 		return checkDuration(strFile, nFileSize);
 		(void)nFileSize;
 	}
 #endif
 
-	UINT checkMediaDuration(IMedia& media, int& nFileSize)
+	UINT checkMediaDuration(IMedia& media, long long& nFileSize)
 	{
 		nFileSize = _openFile(media.GetAbsPath(), media.isXmsc());
 		if (nFileSize <= 0)
@@ -90,29 +90,29 @@ public:
 
 	UINT checkMediaDuration(IMedia& media)
 	{
-		int nFileSize = 0;
+		long long nFileSize = 0;
 		return checkMediaDuration(media, nFileSize);
 		(void)nFileSize;
 	}
 
-	static UINT checkDuration(const wstring& strFile, int& nFileSize)
+	static UINT checkDuration(const wstring& strFile, long long& nFileSize)
 	{
 		return _checkDuration(strFile, false, nFileSize);
 	}
 	static UINT checkDuration(const wstring& strFile)
 	{
-		int nFileSize = 0;
+		long long nFileSize = 0;
 		return checkDuration(strFile, nFileSize);
 		(void)nFileSize;
 	}
 
-	static UINT checkDuration(IMedia& media, int& nFileSize)
+	static UINT checkDuration(IMedia& media, long long& nFileSize)
 	{
         return _checkDuration(media.GetAbsPath(), media.isXmsc(), nFileSize);
 	}
 	static UINT checkDuration(IMedia& media)
 	{
-		int nFileSize = 0;
+		long long nFileSize = 0;
 		return checkDuration(media, nFileSize);
 		(void)nFileSize;
 	}
